+++ src/main/java/com/course/libraryAdmin/controller/LoanAndReturnController.java
@@ -1,4 +1,4 @@
-package com.course.admin.controller;
+package com.course.libraryAdmin.controller;
 
 import com.course.admin.entity.Borrower;
 import com.course.admin.service.ValidateService;
@@ -15,7 +15,7 @@ import org.springframework.web.servlet.ModelAndView;
 import java.util.List;
 
 @RestController
-public class ValidateController {
+public class LoanAndReturnController {
 
     @Autowired
     ValidateService validateService;
@@ -86,5 +86,63 @@ public class ValidateController {
 
         flag = "已借出或已被预约";
         return flag;
+    }
+
+    @RequestMapping(value = "/validateLibraryCode.action")
+    public String validateLibraryCode(@RequestParam("libraryCode") String libraryCode){
+        if(!itemAdminService.isExist(libraryCode)) {
+            return "yes";
+        }
+        else
+            return "no";
+    }
+
+    @RequestMapping(value = "/toWork.action")
+    public ModelAndView toLoan(){
+        ModelAndView modelAndView = new ModelAndView("UserWork");
+        return modelAndView;
+    }
+    @RequestMapping(value = "/toReservation.action")
+    public ModelAndView toReservation(){
+        List<Reservation> list = itemAdminService.findAllReservation();
+        ModelAndView modelAndView = new ModelAndView("UserReservation");
+        modelAndView.addObject("list",list);
+        return modelAndView;
+    }
+
+    @RequestMapping(value = "/loanItem.action")
+    public ModelAndView deleteUndergraduate(@RequestParam("cardNo")String cardNo,
+                                            @RequestParam("libraryCodeLoan")String libraryCode){
+
+        ModelAndView modelAndView = new ModelAndView("redirect:/toWork.action");
+
+        itemAdminService.loanBook(cardNo,libraryCode);
+
+        modelAndView.addObject("cardNo",cardNo);
+        modelAndView.addObject("libraryCodeLoan",libraryCode);
+        return modelAndView;
+    }
+    @RequestMapping(value = "/checkCompensation.action")
+    public String checkCompensation(@RequestParam("libraryCode") String libraryCode){
+        String compensation = itemAdminService.checkCompensation(libraryCode);
+        return compensation;
+    }
+    @RequestMapping(value = "/checkLoseCompensation.action")
+    public String checkLoseCompensation(@RequestParam("libraryCodeLose") String libraryCode,
+                                        @RequestParam("cardNoLose") String cardNoLose){
+        String compensation = itemAdminService.checkLoseCompensation(libraryCode,cardNoLose);
+        return compensation;
+    }
+    @RequestMapping(value = "/returnItem.action")
+    public ModelAndView returnItem(@RequestParam("libraryCodeReturn")String libraryCode){
+        ModelAndView modelAndView = new ModelAndView("redirect:/toWork.action");
+        itemAdminService.returnBook(libraryCode);
+        return modelAndView;
+    }
+    @RequestMapping(value = "/loseItem.action")
+    public ModelAndView loseItem(@RequestParam("libraryCodeLose")String libraryCode){
+        ModelAndView modelAndView = new ModelAndView("redirect:/toWork.action");
+        itemAdminService.addLose(libraryCode);
+        return modelAndView;
     }
 }
