+++ src/test/java/com/course/SystemApplicationTests.java
@@ -2,16 +2,12 @@ package com.course;
 
 import com.course.admin.service.AdminBookService;
 import com.course.borrower.entity.Book;
-import com.course.borrower.repository.BookJPA;
 import com.course.libraryAdmin.service.ItemAdminService;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringRunner;
-
-import java.util.Date;
-import java.util.List;
 
 @RunWith(SpringRunner.class)
 @SpringBootTest
